Eleicoes

+++ Curso4/eleicoes_api/src/services/EleicoesService.js
@@ -1,29 +1,38 @@
-import * as eleicoesRepository from '../repositories/EleicoesRepository.js';
-import * as candidatoEleicaoRepository from '../repositories/CandidatoEleicaoRepository.js';
+import * as EleicoesRepository from "../repositories/EleicoesRepository.js"
+import * as candidatoEleicaoRepository from '../repositories/CandidatosEleicaoRepository.js';
 import * as candidatosService from './CandidatosService.js';
 import AppError from '../errors/AppError.js';
-
 export async function listarEleicoes() {
-    const res = await eleicoesRepository.listarEleicoes();
-    return res;
+    return await EleicoesRepository.listarEleicoes();
 }
 
 export async function obterEleicao(id) {
-    const res = await eleicoesRepository.obterEleicao(id);
+    return await EleicoesRepository.obterEleicao(id);
+}
+export async function resumoEleicao(id){
+    const res = await EleicoesRepository.obterEleicao(id);
+    res.votos = await EleicoesRepository.resumoEleicao(id);
+    res.votos = res.votos.map(el=>{
+        return {
+            id: el.id,
+            nome:el.nome,
+            votos: parseInt(el.votos)
+        }
+    }).sort((a,b) => b.votos - a.votos);
     return res;
 }
 
-export async function resumoEleicao(id) {
-    const res = await eleicoesRepository.obterEleicao(id);
-    res.votos = await eleicoesRepository.resumoEleicao(id);
-    res.votos = res.votos.map(el => {
-        return {
-            id: el.id,
-            nome: el.nome,
-            votos: parseInt(el.votos)
-        }
-    }).sort((a, b) => b.votos - a.votos);
-    return res;
+
+export function criarEleicao(candidato) {
+    return EleicoesRepository.criarEleicao(candidato);
+}
+
+export function atualizarEleicao(candidato) {
+    return EleicoesRepository.atualizarEleicao(candidato)
+}
+
+export function deletarEleicao(id) {
+    return EleicoesRepository.deletarEleicao(id);
 }
 
 export async function adicionarCandidatoEleicao(candidatoId, eleicaoId) {
@@ -40,19 +49,9 @@ export async function adicionarCandidatoEleicao(candidatoId, eleicaoId) {
 }
 
 export async function removerCandidatoEleicao(candidatoId, eleicaoId) {
+    const candidato = await candidatosService.obterCandidato(candidatoId);
+    if (!candidato) {
+        throw new AppError("Candidato não encontrado", 404);
+    }
     return await candidatoEleicaoRepository.removerCandidatoEleicao(candidatoId, eleicaoId);
-}
-
-export async function criarEleicao(eleicao) {
-    const result = await eleicoesRepository.adicionarEleicao(eleicao);
-    return result;
-}
-
-export function atualizarEleicao(id, eleicao) {
-    const result = eleicoesRepository.atualizarEleicao(id, eleicao);
-    return result;
-}
-
-export async function deletarEleicao(id) {
-    await eleicoesRepository.deletarEleicao(id);
 }
